feat: use node_module browser-remaps instead of local package.json
