+++ .eslintignore
@@ -1,6 +1,7 @@
 .vscode
 .nyc_output
 node_modules
+!src/node_modules
 coverage
 dist
 __snapshots__
