+++ .gitignore
@@ -7,6 +7,7 @@
 
 # NPM
 node_modules
+!src/node_modules
 npm-debug.log
 
 # Build
