+++ .prettierignore
@@ -3,6 +3,7 @@
 package.json
 package-lock.json
 node_modules
+!src/node_modules
 coverage
 dist
 __snapshots__
