+++ src/components/micro-frame-slot/index.marko
@@ -1,4 +1,4 @@
-import slotComponent from "./component";
+import slotComponent from "@internal/micro-frame-slot-component";
 
 // We have a proxy to the top level component
 // so that we can swap it out using `package.json#browser`.
