+++ src/components/micro-frame/index.marko
@@ -1,4 +1,4 @@
-import streamedComponent from "./component";
+import streamedComponent from "@internal/micro-frame-component";
 
 // We have a proxy to the top level component
 // so that we can swap it out using `package.json#browser`.
