--- src/components/stream-source/component/package.json
@@ -1,4 +1,0 @@
-{
-  "main": "./node.marko",
-  "browser": "./web.marko"
-}
