+++ src/components/stream-source/index.marko
@@ -1,4 +1,4 @@
-import streamSourceComponent from "./component";
+import streamSourceComponent from "@internal/stream-source-component";
 
 // We have a proxy to the top level component
 // so that we can swap it out using `package.json#browser`.
